main: Generate our own options text for the usage message.
As noted in the previous commit, the three-column message we're
getting, which is provided by System.Console.GetOpt, is never going to
fit into 80 columns.

Instead, set up a two-column format with the options on the left and
the descriptions on the right, use a fixed column size, and make the
layout smarter so the right column adjusts around stuff that protrudes
out of the left column.

Alas, the results with this are 81 columns wide, so a bit more editing
will still be needed.