CI: Don't run code coverage job on forks
The code coverage job uses `deploy-pages` to publish the code coverage results,
but this requires `GITHUB_TOKEN` permissions from someone within the GaloisInc
GitHub organization. As such, this job cannot reasonably work on a fork of the
`saw-script` repo, so we disable this job if a fork is detected.

Fixes #2216.