saw-user-manual/structure-of-sawscript: Move first simple example.
As @smithdtyler rightfully points out, this was way later in the manual than it
ought to have been, so this commit moves it up where they suggest, and tweaks
the wording slightly to better fit in.

Despite that tweak, the same "under construction" message used previously is
repeated here to emphasize the content is subject to change.