saw-user-manual/appendices: Make separate directory for appendices.
Previously, for the first appendix, we made it a top-level document of the user
manual.
Since there are a number of other resources we may wish to add to what we
consider to be 'appendices' of the manual (e.g. a glossary, a command reference,
a full SAWScript language reference), this commit preemptively adds some
additional hierarchical structure to the manual to better organize these
resources, even if for the time being there is only one.

Note that this commit includes appropriate re-titling/re-naming of the one
existing appendix, to avoid excessive or redundant use of the word 'appendix' in
titles etc.