appendices/glossary: Add an (empty) glossary to the appendices.
This document will be the home of definitions of terminology used throughout the
manual that users new-ish to verification, and completely new to SAW, are
unlikely to know.

I decided to leave the glossary empty for the time being, but adding new terms
is very simple:

Definition lists that have a {.glossary} attribute will be able to be
cross-referenced using the {term} role.
So, to add terms, under an appropriate {.glossary}, write:

```markdown
My term
: The definition of my term
```

Then, anywhere else in the document tree, this definition can be referenced
with:

```markdown
{term}`my term`
```

Note that capitalization in the reference is irrelevant.
Sphinx will emit warnings for anything annotated {term} that does not have an
entry in a {.glossary}.