saw-user-manual/appendices: Properly format REPL reference.
In particular, use proper definition list formatting.
This allows the lists to be turned into glossaries, if desired; as an example,
this commit makes the definition list of environment variables a {.glossary},
and adds some references to these terms to the text where they are mentioned (at
least, the first use -- this isn't a wiki, after all).