saw-user-manual: Place proofs about terms after term transformation.
As written, this is a more logical ordering of these manual sections.
Most importantly, the proof-about-terms document references ideas introduced in
the transforming-term-values document; thus, the latter is prerequisite.

Later updates to the manual may revisit this, but significant changes to the
text will be necessary to discuss proofs before transformations.