Bump what4 submodule (and friends) to bring in GaloisInc/what4#277 changes
This bumps the `what4` submodule to bring in the `BoolMap`/`ConjMap`-related
changes from GaloisInc/what4#277. (In turn, this requires bumping the
`crucible` submodule to bring in the changes from GaloisInc/crucible#1306, as
well as the `macaw` submodule to bring in the changes from
GaloisInc/macaw#462). This requires some minor changes in `saw-core-what4` to
accommodate.

Fixes #2223.