appendices: Replace REPL glossary entries with explicit references.
This is to avoid undesirable styling of these glossary items in documents (in
particular, glossary terms render italicized/emphasized, and this is not
appropriate for things like `PATH` and other REPL terms).

The approach sacrifices the inline convenience of glossary terms, which allows
for freedoms like different capitalization, to achieve control over styling -
since this explicit reference approach can be used as any other link.