README: Fix broken documentation links.
In particular, the link to the manual was outdated.

This commit also changes the other tutorial links to instead point to the
in-repo PDFs, as these are more up-to-date than the resources at
tools.galois.com (and, the resources at this latter site are ephemeral / going
to be removed as soon as docs are deployed to GH Pages).