Add haddock to the CI build.
Run it for the whole cabal project, and for each GHC version, but only
on the Linux builds.

Apparently the way to generate documentation for the packages you
asked for and not everything under the sun as well is to pass
"--disable-documentation".

There's a "cabal haddock-project" to do the whole project
automatically, but it doesn't accept --disable-documentation, so we
need to use "cabal haddock" and list the packages we want the haddocks
for. Put that in ci.sh as the list is too big to stuff into ci.yml.