Add -B / --batch to execute REPL commands from a file.
It turns out the REPL already knows how to execute from a file,
there's just been no way to get to the functionality. Hook it up,
since we want to be able to test :search and other :-commands.

Unfortunately the option is -B rather than matching Cryptol's -b, but
our -b is already taken.