search: simplify the Ord instance for Candidate
It had been first calling Eq, I think probably to make sure that it
returns Eq for the cases where the types are equal according to
matchExact. But that causes it in general to do two recursive
examinations of the objects, which is wasteful, and also isn't
particularly effective because it could also return Eq for other
things and that would also be wrong.

Instead, document that the arbitrary ordering on types has to be
consistent with matchExact. Can't prove it in this environment,
otherwise I'd do that...