llvm: Print more when an assertion doesn't match.
There was code here to print the terms involved, but it was commented
out; uncomment it and make it work.

This produces fairly illegible saw-core terms but it does give you
some hint of _which_ assertion failed, instead of just saying "Literal
equality postcondition" or "Literal equality precondition".

Also, it seems that showTerm generates multiline strings that expect
to start at the beginning of the line, so accomodate that. This
doesn't produce particularly great formatting, but we'll need more
printing infrastructure to do better. (FUTURE)

Add a test that visits this message via having an incorrect
llvm_return. Create a test_llvm_errors subtree for such tests,
and use test-and-diff.sh.

Starting point on #277.