interpreter: fake a better filename to read the builtin type signatures
Use "<type of ...>" instead of "internal". Now we get e.g.
   <type of true>:1:1-1:5: The type Bool arises from this type annotation
which is by no means ideal but can at least be followed.

Doing better than this will be hard though because we stuff the type
signatures through the parser rather than consing them up directly.
Either we have to stop doing that, or postprocess the positions
somehow, to get something like "Type of 2nd argument of concat" as the
position. Which is really what we want.

However, this'll be good enough for now.

Specifically pursuant to #457, but has broad applicability.