tidy: Fix the name of crux-mir-comp's test suite.
Calling it just "test" lacks a certain something but also conflicts
with one in crucible, and cabal isn't super happy about that.

There don't appear to be any direct references to it that need to be
updated, but I might have missed something as it's not a good thing
to search for.