reorg: Move Version.hs to a new saw-version library.
Move GitRev.hs (and its generator script) too.

This will allow saw-remote-api access to the version info.

Renames only; build changes next.

I had been hoping that we'd be able to put the saw-version sources in
src/SAWVersion (so we'd have one library in src/SAWVersion and the
other in src/SAWScript), but alas, sharing source directories like
this confuses cabal and/or ghc and leads to spurious link failures.

Failing that, I'd also been hoping to just have saw-version/SAWVersion
without an extra "src" layer, but while that's fine for saw-version,
there are several libraries that have additional pieces and need a
"src". After discussion we concluded it would be better for everything
to have the same structure.

So put things in saw-version/src/SAWVersion, and we'll just have to
deal with the two extra vacuous layers of directory tree.