reorg: Move Version.hs to a new saw-version library.
Move GitRev.hs (and its generator script) too.

This will allow saw-remote-api access to the version info.

Renames only; build changes next.

I had been hoping that we'd be able to put the saw-version sources in
src/SAWVersion (so we'd have one library in src/SAWVersion and the
other in src/SAWScript), but alas, sharing source directories like
this confuses cabal and/or ghc and leads to spurious link failures.

So we are stuck having saw-version/SAWVersion instead.