savegitinfo: Don't regen GitRev.hs if .git isn't there.
It obviously won't do anything useful in that case, and if .git has
disappeared on purpose (e.g. by doing git archive, copying into a
container or VM, etc.) we want to keep the previous info rather than
clobber it with a bunch of Nothing.

However, if there isn't a GitRev.hs file, generate one anyway.