.github/workflows/ci.yml: Add build-docs job.
This commit adds the start of a documentation-building CI job.

After coverage is computer, this job attempts to read the Python environment
from cache, run the Sphinx build, collect all HPC HTML, and copy these files to
the Sphinx HTML output so they will be included in the eventual deployment to
GitHub pages.