.github/workflows/ci.yml: Update the coverage job.
This commit removes the collection of HPC coverage files (added to the
build-docs job in 286c5c916), as well as deployment to GitHub pages, from the
coverage job.

Because these changes remove all GitHub Pages interaction from the job, this
commit also removes the unnecessary if: and permissions: properties.
These are not required with no GitHub Pages deployments in the job.