.github/workflows/ci.yml: Add job to update documentation versions.
Recall that 4f40c6381 added a versions pane to the HTML templates, as well as
JavaScript to update this pane with the currently-available documentation
versions, and relevant links.

This commit adds the corresponding CI job, which checks out the gh-pages
deployment and runs a Python script that generates a JSON file listing the
documentation versions that have been deployed.