reorg: Create a saw-central library.
Seed it with Options, Position, and Utils.

These are the bottom-level modules in the current saw-script library
and all of them are (for the time being at least) needed by the Python
interface.

The idea with saw-central is that eventually the saw-script library
will contain the saw-script interpreter, and saw-central will contain
the guts of SAW that are used by both the saw-script interpreter and
the remote API. And then the remote API won't need to half sit on top
of the interpreter and can be a first-class citizen.

saw-central is the best name for it that I've managed to come up with,
given that saw-core is already taken for something else.

This commit contains renames only for git safety. Module and build
changes to follow.