.github/workflows/ci.yml: Upload coverage separately.
Previously, we copied coverage reports _per version of the docs_.

This doesn't make sense given how the coverage report gathering process works,
and results in a large amount of duplicated data on the gh-pages branch.

This commit adds a step to instead separately deploy coverage reports to the
coverage/ directory on gh-pages, and no longer duplicate them per documentation
version.