.github/workflows/ci.yml: Don't include 'coverage' in doc versions.
While this ends up 'working' (in that clicking to change to the coverage version
of the docs instead takes you to the coverage report home), it doesn't make
sense given the function of the version-switching pane.

This commit adds an additional condition to the filtered directory names,
explicitly excluding a directory named 'coverage'.