doc/conf.py: Use rev-parse instead of symbolic-ref to get branch.
git symbolic-ref --short HEAD exits with an error if run while in a detached
HEAD state (which is the case in a GitHub action)

git rev-parse, on the other hand, does not -- it returns "HEAD" in that state.

This is sufficient for CI, as this particular setting in conf.py is not critical
to the page deployments behaving as expected.