.github/workflows/ci.yml: Split doc build and deployment jobs.
In particular, to always build the documentation, but restore the check that
this is not being run in a fork (due to permissions required to deploy to GitHub
Pages).

This commit also restores this same check to the coverage job, which has had its
own deployment to GitHub Pages restored.