doc/conf: Better handling of current_version.
GitHub actions end up checking out the repository in a detached HEAD state; this
means that in the HTML documentation build, current_version will be set to the
name HEAD.

I previously thought this was OK, in that it wouldn't happen for master/version
tags.
But this isn't the case, obviously, as this becomes a permanent aspect of the
render no matter what.

This commit attempts to resolve this by introducing a BRANCH_NAME variable to
the build environment, which in CI is computed as the head_ref (i.e. source
branch for a pull_request), or ref_name (the GitHub branch/tag name).
This allows us to avoid the issues with running the git commands in conf.py
while in a detached HEAD state.

BRANCH_NAME takes priority over those commands; if available, they are not used
at all.