.github: Add similar handling to previous commit to directory names.
This will result in us publishing documentation for all PR source branches, but
this isn't necessarily a bad thing for the time being / while working out kinks
in the documentation build and deployment.

This uses the same BRANCH_NAME environment variable as the previous commit,
where similar handling was added for the documentation build step.