.github/workflows/ci.yml: Run deploy-docs on push as well.
Previously, because of the condition on it, the deploy-docs job would _only_ run
for pull_request events on non-forks.
This commit updates the condition to allow the job to run when the event is
'push', meaning we will also deploy for merges to `master`, `release-*`
branches, and `vM.m`/`vM.m.p` tags.