.github/workflows/ci.yml: Don't run deploy-docs on release-*.
This prevents an awkward "partial run" for such branches -- previously, this
_job_ would run for release-* branches, but the _step_ pushing documentation to
the gh-pages branch would be skipped (due to the destination directory
determination returning an empty string).

Since we already restrict the CI workflow to pushes to tags / master / release
branches, it is easiest to specify the condition as a denial of branches
starting with release- than acceptance of master / all version tags.