tidy: Consolidate LICENSE files.
They are all BSD 3-clause, but we had multiple variants with different
copyright years, different formatting, and slightly different wording
in the use-of-name clause.

Cabal seems to want each subpackage to continue to have its own copy
(if you point the subpackages at the top-level copy it starts
complaining) but they can at least be identical.

Merge the copyright years and set the end year of the range to 2025.