doc/saw-user-manual: Remove references to addsimp' and addsimps'.
This resolves #1814.
This includes some lightweight reordering and rewriting of the text, to avoid
broken examples / more naturally move into the introduction to the Theorem type.

As part of this commit, the addimp' and addsimps' functions have been added to
the "Deprecated Items" Appendix, with a brief description (based on a paragraph
that was removed elsewhere) of _why_ the items were deprecated.