pyproject.toml: Try the explicit path to the cryptol-remote-api python.
(the new version hasn't been published yet so it can't be downloaded)

If this gives a working test run, I'll revert it again before merging,
and we can merge in a theoretically broken state because that
publishing will happen at the same time the saw-remote-api python gets
published.

But I would really rather see a success in the remote API tests before
merging this branch, because if there is some kind of problem we want
to know about it _before_ things start getting published.