CI: Allow the deploy-docs job to run on release-** branches.
This is a workaround for not getting runs for tag pushes. Not clear
why; they just aren't happening.

Also, adjust save_doc_publish_dir to produce nonempty output for
release branches. Otherwise the actual deploy step gets skipped. Note
in ci.yml that this is necessary so it doesn't get forgotten next
time there are adjustments.