CI: fix the on: push: tags: pattern.
It's a "glob" not a regular expression and doesn't support grouping
with parens.

It does apparently support regexp-style ? and + operators, so the
existing pattern wasn't as completely wrong as for shell globs.

Fortunately you can have multiple cases which are OR'd so we don't
actually need the grouping.

This is almost certainly why we haven't been getting tag runs.