s2ntests: Tell the s2n infrastructure not to install saw.
This prevents it from downloading and trying to run an ancient binary
that no longer runs on Ubuntu 24.04.

Downside of this is that this also turns off fetching yices and z3, so
instead copy in the yices and z3 we already have. This may or may not
work since they're probably quite a bit newer than the ones expected,
but if they do it's probably for the best.

This also may not be the way we want to do this, but if it works it
avoids needing to change the s2n tree.