s2ntests: use ubuntu clang-14 instead of an LLVM 10 binary
The binary distribution seems to be linked against ncurses 5, which is
no longer available in ubuntu 24.04.

Why a compiler needs to be linked against curses, IDK.

Applies to the awslc and blst containers, but not (yet?) the default
s2n container as that uses a different clang that might or might not
be runnable.

Provisional.