build.sh: move to a model where build.sh takes an optional target arg
(like you would run make)

The discussion in #2063 has more or less converged on this much.

This commit adds the ability to do "build.sh submodules" (which
does git submodule update --init, because the build already does
that by default) and "build.sh clean".

The other bells and whistles discussed in #2063 are probably coming
fairly soon but this much is needed right away.