CI: Run "build.sh gitrev" before building the docker images.
This will generate a GitRev.hs with the current git state that then
gets copied into the images. It will then not be clobbered inside the
image because savegitinfo.sh won't overwrite an existing GitRev.hs if
no .git tree is present.

This fixes the known problem where the docker images don't have git
information. There doesn't seem to be an open issue for it though.