Add a copy of SAWCore.hs from crucible-saw.
This will be SAWScript.Crucible.CrucibleSAW.SAWCore in the module
namespace, at least for now. (The old module path was
Lang.Crucible.Backend.SAWCore.)

This is a verbatim copy of the file from crucible-saw as it existed
when it was deleted from the crucible tree two months ago. It won't
build without updating its module header, but I'm going to commit that
separately for ease of possible later rearrangements.