CrucibleSAW: cope with OnlineBackend changes
Interface changes have assimilated some of the code that was here into
the Crucible OnlineBackend interface. So we don't need that any
more. However, those changes also make the initialization we need to
do circular. The only viable way to break that cycle without upstream
interface changes is to set the initial online state to Nothing and
then update it after we've constructed the objects. This means it will
always be Just at runtime, so fake out the prior field accessor to
unwrap the Maybe and panic if it's Nothing, which it should never be.

XXX: This is using the wrong panic interface; clean that up later if
XXX: this code sticks around.

XXX: This is a hack. If this code sticks around, we should look into
XXX: upstream interface changes to avoid the cyclic initialization.