saw-core-what4: Improve the failure prints in rebuildTerm.
It will now say "saw-core-what4: rebuildTerm: cannot handle foo"
instead of just "rebuildTerm foo", and it will attempt to Show the
Value (there's no prettyprinter that I can find) although this won't
help much in many of the cases.

This is not "fixed" but it's better than the status quo.