Disable heapster examples again.
This branch (Fix examples after EnTree #1827) fixed some of them, but
got left unmerged and in the meantime got broken again by #2017 (which
rearranged the SpecM monad without updating the proof automation on
the Coq side, as tracked in #2026). As best I can tell we want these
changes, but further work will be needed before the examples are
buildable again.

This commit disables the proofs that had been re-enabled on this
branch. There are also others that were still broken in this branch's
original timeframe and haven't been fixed since. These all depends on
SpecMExtra.v in the Coq support, which is currently disabled.