Allow SAWScript lexer to fail with a message
Invalid strings (bad escapes, multiline) are explicitly disallowed.
Numeric literal prefixes are always a prefix, and 0b2 is now an error
rather than a number token and an identifier.

Now that the lexer can fail, it made sense to change the type to only
return tokens on success. It also eliminates the parse error message on
every lexer failure. I considered attempting to recover but I don't
think it's appropriate for lexing.