MIR backend: Support nightly-2025-02-16
This updates SAW and `crux-mir-comp` in order to support a more recent Rust
nightly (`nightly-2025-02-16`). This piggybacks off of the corresponding
`mir-json` changes in https://github.com/GaloisInc/mir-json/pull/98 (a fix for
https://github.com/GaloisInc/mir-json/issues/78), as well as the
`{crucible,crux}-mir` changes in
https://github.com/GaloisInc/crucible/pull/1396.

Most of the interesting changes are on the Crucible side in
https://github.com/GaloisInc/crucible/pull/1396. The code in SAW that had to
change:

* `mirRef_offsetSim` no longer takes a `TypeRepr` argument, so some code in
  `crux-mir-comp` had to be adjusted accordingly.

* In order to look up the `DefId`s for lang items (e.g., `$lang::Option`), the
  `DefId` lookup logic now takes into account the new `lang_items` table in a
  MIR JSON file.

This also updates all of the vendored-in `*.linked-mir.json` files and
`crux-mir-comp` test cases to use updated line number information.