saw-core: Rename the untyped Term to UTerm.
(plus some related bits)

It's really confusing having two different Term types (from the
untyped and typed saw-core ASTs) especially since the saw-core library
doesn't currently have any particular structure that would lead us to
believe the untyped version is private to the library.

And it turns out that while it's actually pretty close, it does leak
out in one place. That should get corrected (doing so should be
straightforward, but requires the pending error reporting cleanup) and
eventually the saw-core library should be rearranged so you _can't_
refer to the untyped AST from outside.

This change only renames a couple of the types (and in particular not
their corresponding data constructors), which is not entirely optimal
but will do for now. Can do more cosmetic cleanup later.