saw-core parser: don't "error" on a missing module name
Previously, if you wrote "module where" at the top of a SAWCore file
(instead of "module Foo where"), it was accepted by the grammar and
then called Haskell "error" to bail out.

Fix the grammar so this is a parse error, and change the error call
involved to a panic.

This degrades the message printed a bit, maybe, since there's no
longer an explicit "empty module name" message; but it now includes
the offending source position, which is quite a bit more important.

Add a test.

Note that the source position prints as rubbish; that's a different
problem (#2140)