interpreter: Clean up cases where we insert vars with possibly no type.
There were two sources of these: one was llvm_compositional_extract,
which absolutely did insert a variable with no type. As best I can
tell, the (saw-script level, not Cryptol level) type should just be
"Term", so run with that.

The other was a bunch of plumbing in the interpreter which could only
actually send Nothing down if executed on terms that hadn't been
typechecked. Since we are no longer doing that, panic instead further
up if there's no type to be had. This can be cleaned up more later,
but I only want to do so much at once in case something unexpected
pops up.