SAWScript AST: rename "PrettyPrint" to "PrettyPrintPrec".
As noted in #950 it is odd to have this class with the same naming
and different signature (it includes a precedence argument) as the
normal prettyprinter.

I'm about to kick this out of here where it doesn't belong; renaming
it first avoids name conflicts.

Also it's a little sad that we define this and then don't actually use
it for expressions, only for types. That should probably get fixed up
sometime.