Pretty: collect prettyprinting infrastructure into a new SAWSupport.Pretty.
In the absence of anywhere to put support material it's been stuffed
in anywhere, most of it being either (a) embedded in saw-core's
prettyprinting, (b) stuck in saw-script's AST module, or (c) randomly
dumped into Value.hs. Pull from all of these.

There are a few bits from Value.hs that are Show infrastructure.
These should go away in the future; the things that use them should be
using the prettyprinter interface instead.