CI: adjust ordering in build().
Run "git status --porcelain", whose purpose is to check for garbage
in the tree, before we start on the configure and build.

Move the assignment of "pkgs", which is part of the build step,
after the configure step, instead of doing it in the middle.

Also document that the adjustments to cabal.project.local must come
_after_ running cabal configure. One might reasonably think that cabal
configure would read cabal.project.local as part of the configuration
process, and reads it again if it changes; but that's not the case.
Rather, it _creates_ it, so any adjustments to it must happen
afterwards.