Kill off depthPPOpts and ppTermDepth.
Restore the place that had been using it but had been commented out.

Limit the depth correctly, by clamping the current depth setting
(instead of silently ignoring preexisting settings) ... except that
we can't currently get the real current printing options, so start
from the default ones for now. But this at least moves the problem
(an instance of #2353) up a layer.