build.sh: allow setting SAW_SUPPRESS_GITREV to avoid updating GitRev.hs.
Updating GitRev.hs incurs a rebuild of the SAWCore prelude, in order
to make sure that in the event that doing so generates a panic, the
panic includes the correct git version data. While a nice idea, this
is annoying because rebuilding the SAWCore prelude is quite slow.

This is a hack that should only be used when actively doing
development (such that if something blows up on you, you are sure
the version that blew up is whatever you just built) and if anyone
knows a better way to cope with this problem please let me know.

Note that marking the git version string noinline does successfully
prevent git state changes from recompiling everything that directly
or indirectly refers to "panic" (which is basically the whole system)
but the SAWCore prelude is built with Template Haskell so it's very
conservative about rebuilding anytime anything changes.