interpreter: kick mutable material out of TopLevelRO.
TopLevelRO uses MonadReader and you can update the state in a
MonadReader as long as you do it let-binding style by wrapping actions
in a different state withFoo-style. There is nothing wrong with that
per se; however, the fact that TopLevelRO wasn't actually readonly has
been
   - confusing;
   - incorrect.

Several things manipulate the TopLevelRW and think they're handling
the complete interpreter state. This may or may not actually be true
in any given situation.

Then, also, we've been using this with update flow that is _not_
let-binding style. You can wedge it into that format by passing
monadic actions around instead of just executing them. Sort of. But
doing so makes a difficult-to-follow and fragile mess.

This change moves roPosition, roStackTrace, and roLocalEnv to
rwPosition, rwStackTrace, and rwLocalEnv respectively.

XXX: It is also super ugly that which state is readonly and which is
XXX: not gets exposed in the public interfaces.