getW4Position: remove outdated/lying comment and simplify slightly
The comment seemed to be claiming that because the FromValue instance
for CrucibleSetup wraps the LHS of any VBinds it sees in the VBinds'
own positions that fetching the position out of the CrucibleSetup
instead of the interpreter state would be more accurate. That might
have been true at some point, but in the recent past there's only been
one position to fetch, out of the interpreter state inside the
CrucibleSetup, which is also the one updated in the FromValue
instance in question.