interpreter: rename localOptions to withOptions for clarity.
Note that this is still a remaining case of updating the "readonly"
part of the interpreter state: the output-tracing functionality
(--output-locations command-line option) works by wrapping interpreter
monad closures in an updated Options that changes "printOutFn".
Having printing work via a closure hung in the command-line options
record is not reasonable and fixing it is part of a different problem;
in the meantime the way "printOutFn" gets manipulated is a nice set of
deck chairs on RMS Titanic, so it's getting left alone for now.

(I'm just going to do this rename because it's next to the other
related things I've been doing.)