SAWScript: remove unused AST elements CryptolSetup and CrucibleSetup.
CryptolSetup existed as its own monad type (like LLVMSetup) but there
were no available actions in this monad. So you could write vacuous
declarations like "let three : CryptolSetup Int = return 3;" but not
do anything with them.

CrucibleSetup did too, but any reference in the input was (and still
is) intercepted in the parser and turned into LLVMSetup, so it wasn't
accessible at all.

G/C the lot.

The result of this is that if anyone actually has that declaration
"three" it will still be accepted: it'll treat "CryptolSetup" as a
free type variable and produce an object of polymorphic monad type
"{m} m Int". This is not entirely ideal, but the existence of any
such declarations in the field is unlikely and they still don't
do anything particularly useful.