saw-script interpreter: avoid overlapping assignments of global variables
The move of LocalEnv from TopLevelRO to TopLevelRW caused a case where
TopLevelRO was updated inside an update to TopLevelRW to become an
update of TopLevelRW inside another update of TopLevelRW. This doesn't
work so well.

The two assignments involved were in different source modules, so it
took some finding.

Fix it by eliminating the additional assigment of LocalEnv and instead
passing the LocalEnv in question as an argument, which is probably how
the code should have been written in the first place. Updating globals
to call functions that use them as inputs is very BASIC.

This restores the dynamic rebind behavior described in #1646, closing

It isn't entirely clear why this only broke the dynamic rebind
behavior and didn't, say, lose the rebind entirely or have other more
noticeable effects, other than that the code is a maze.