Remove the "callcc" SAWScript primitive.
It was implemented using a Haskell-level callcc in a shallow
embedding, on the assumption that you can just teleport around without
breaking the interpreter state, or the saw-core state, or various
other things. You can't.

If some reason appears, which seems unlikely, we can do a real
implementation of callcc in the future.

Note that I'm actually removing it, not marking it deprecated. It was
marked experimental, described as "highly experimental", and as far as
we know, it's only been used once as part of a debugging hack.