saw-script interpreter: unhook recordProof/returnProof from IsValue.
There are four kinds of proofs we can record (Theorem, and proved
method specs from the Crucible LLVM, JVM, and MIR backends) and they
go into Value, so IsValue is convenient. However, IsValue and
FromValue are highly tangled with the interpreter logic (among other
things, critical parts of the eval logic are pasted into various
FromValue instances) so this convenience comes at an excessive cost.

Also, probably we're better off not having a way to record "proofs"
that are other random instances of Value, such as VBool False.