interpreter: Remove the overlapping IsValue/FromValue instances for String.
The interpreter now deals only in Text.

Update some test reference outputs for some print changes.

Take the opportunity to textify a bunch of the crucible-llvm code.

The disadvantage of doing this is that because FilePath is still
String, there's now a substantial number of one-line thunks that
manually unpack Text to FilePath before calling into the rest of the
system. This is the kind of thing typeclasses are supposed to handle,
so we might want at least the FromValue instance back later after
other things have been cleaned out.