interpreter: cut back on stuffing closures into values.
When we get a saw-script level lambda, make that a value instead of
closing a copy of the interpreter in with it.

The practical effect of this is limited because the way stack traces
are done wraps those immediately in another closure. Going to need to
change that around later.