Fix saw-core-{sbv,what4} buglets involving Vec-related model values
Neither the SBV nor What4 backends of SAWCore were particularly robust when
tasked with producing models for `Vec`-related values (e.g., values
corresponding to the Cryptol types `[2][8]` or `[0][8]`) when using commands
such as `sat` or `sat_print`. In particular, the SBV backend would crash when
producing an empty `Vec`, and the What4 backend would be liable to crash when
trying to produce any `Vec` values at all. This led to the issues seen in #1094
(for the SBV backend) and #2464 (for the What4 backend).

Fortunately, both issues can be fixed in a relatively simple fashion: simply
store the element type (as a `FirstOrderType`) in the `VecLabel` case of each
backend's `Labeler` data type.

Fixes #1094. Fixes #2464.