interpreter: fix handling of do-blocks.
The prior behavior was to execute into do-blocks until the first bind,
then defer execution using a chain of VBind value thunks with copies
of the interpreter closed into them. This caused a lot of unnecessary
consing; however, worse than that it caused strange eval behavior.
For example,
   let f = do { let x = undefined; return 3; };
crashes immediately, but
   let f = do { _ <- return (); let x = undefined; return 3; };
crashes only when the resulting monadic value is executed.

To fix this, introduce a VDo value, containing abstract syntax rather
than values, and when we hit a do-block when executing purely, return it
immediately as a VDo value. Then, in the same places that process VBind,
process VDo by executing the statements in it.

This changes the internals of the interpreter slightly such that
expressions are always evaluated purely and statements are always
evaluated in a monad context; it is thus necessary when executing
statements to force any VDo values that appear. (The typechecker
ensures that they'll be do-blocks in the correct monad.)

Unfortunately, we can't get rid of VBind yet. There is one thing that
still needs it: the "for" builtin. It can't use VDo: as a builtin, it
is handed only values, not abstract syntax, and for the time being we
can't lift arbitrary values to abstract syntax. The best bet is
probably to rewrite it in SAWScript and stick it in the prelude we
don't yet have (see #253). Rename VBind to VBindOnce and note all
this.

One further caution: this commit removes part of the stack trace in
the reference output for test_sawscript_builtins/fail2. This is
because the current scheme for doing stack traces can't cope with
the new VDo, and I'm not inclined to garbage things up to make it.

I'm going to add some chewing gum and baling wire to bring the stack
trace entry back in the next commit, but I expect to be able to revert
the mess in the near future. :-)